+++ src/popup/data.js
@@ -1,6 +1,6 @@
 const data = [
   {
-    url: /chess\.com/,
+    url: /^chess\.com/,
     quotaRules: [
       {
         daysOfweek: [0, 1, 2, 3],
@@ -8,9 +8,9 @@ const data = [
           {
             startTime: '00:00',
             endTime: '23:59',
-            quota: 30
-          }
-        ]
+            quota: 30,
+          },
+        ],
       },
       {
         daysOfweek: [2, 3, 4],
@@ -18,16 +18,16 @@ const data = [
           {
             startTime: '00:00',
             endTime: '12:00',
-            quota: 30
+            quota: 30,
           },
           {
             startTime: '13:00',
             endTime: '20:00',
-            quota: 30
-          }
-        ]
-      }
-    ]
+            quota: 30,
+          },
+        ],
+      },
+    ],
   },
   {
     categoryId: '20', // gaming
@@ -39,12 +39,12 @@ const data = [
           {
             startTime: '00:00',
             endTime: '23:59',
-            quota: 30
-          }
-        ]
-      }
-    ]
-  }
+            quota: 30,
+          },
+        ],
+      },
+    ],
+  },
 ]
 
 export default data