+++ src/popup/utils/index.ts
@@ -1,7 +1,9 @@
+import { RuleObject } from '../types'
+
 /* 
 @input: rule object
-@output: the title of the rule if url
+@output: the title of the rule 
 */
-export const getRuleTitle = rule => {
-  if (rule.url) return `${rule.url}`.match(/\w+/)
+export const getRuleTitle = (rule: RuleObject) => {
+  if (rule.url) return `Url: ${rule.url}`.match(/[^\\\/\^\$]/gi)
 }