+++ CHANGELOG.md
@@ -2,8 +2,24 @@
 
 - added data file
 
+- <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons" /> 
+  for material fonts
+
+- <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap" />
+    for roboto font
+
+-removed fontawesome CDN , going to use Material Icons
+
+## npm
+
+- npm install @material-ui/core @material-ui/icons
+
 ## src/popup
 
-- added RulePage
-- added utils
-- added types
+- added two tabs in RulePage (Rules, Stats)
+
+- added some styling to RulePage
+
+- Changes header to Functional Component
+
+- Changed settings cog to Material icon